wip: split up runtime-api and rpc-api
