+++ node/Cargo.toml
@@ -88,9 +88,10 @@ xcm = { git = "https://github.com/paritytech/polkadot.git", branch = "release-v0
 
 # Self dependencies
 calamari-runtime = { path = '../runtime/calamari' }
-manta-runtime = { path = '../runtime/manta' }
 dolphin-runtime = { path = '../runtime/dolphin' }
 manta-primitives = { path = '../primitives' }
+manta-runtime = { path = '../runtime/manta' }
+pallet-manta-pay = { path = '../pallets/manta-pay', features = ["rpc"] }
 
 [build-dependencies]
 substrate-build-script-utils = { git = 'https://github.com/paritytech/substrate.git', branch = "polkadot-v0.9.16" }
