+++ pallets/manta-pay/Cargo.toml
@@ -28,6 +28,11 @@ rpc = [
 	"jsonrpc-core",
 	"jsonrpc-core-client",
 	"jsonrpc-derive",
+	"serde",
+]
+
+# Runtime API
+runtime = [
 	"sp-api",
 	"sp-blockchain",
 ]
@@ -38,6 +43,13 @@ runtime-benchmarks = [
 	"frame-support/runtime-benchmarks",
 	"frame-system/runtime-benchmarks",
 	"manta-primitives/runtime-benchmarks",
+]
+
+# Serde Serialization
+serde = [
+	"manta-pay/serde",
+	"manta-util/serde",
+	"manta-util/serde_with",
 ]
 
 # Standard Library
@@ -63,18 +75,18 @@ precompute-coins = [
 [dependencies]
 # utils
 anyhow = { version = "1.0.55", optional = true }
-indoc = { version = "1.0.3", default-features = false, optional = true}
-rand = { version = "0.8.4", default-features = false, optional = true }
+indoc = { version = "1.0.3", optional = true, default-features = false }
+rand = { version = "0.8.4", optional = true, default-features = false }
 tempfile = { version = "3.3.0", optional = true }
 
 # substrate dependencies
-frame-benchmarking = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.16", default-features = false, optional = true }
+frame-benchmarking = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.16", optional = true, default-features = false }
 frame-support = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.16", default-features = false }
 frame-system = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.16", default-features = false }
 scale-codec = { package = "parity-scale-codec", version = "2.3.1", default-features = false, features = ["derive", "max-encoded-len"] }
 scale-info = { version = "1.0.0", default-features = false, features = ["derive"] }
-sp-api = { git = 'https://github.com/paritytech/substrate.git', optional = true, branch = "polkadot-v0.9.16", default-features = false }
-sp-blockchain = { git = 'https://github.com/paritytech/substrate.git', optional = true, branch = "polkadot-v0.9.16", default-features = false }
+sp-api = { git = 'https://github.com/paritytech/substrate.git', branch = "polkadot-v0.9.16", optional = true, default-features = false }
+sp-blockchain = { git = 'https://github.com/paritytech/substrate.git', branch = "polkadot-v0.9.16", optional = true, default-features = false }
 sp-runtime = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.16", default-features = false }
 sp-std = { git = 'https://github.com/paritytech/substrate.git', branch = "polkadot-v0.9.16", default-features = false }
 
@@ -84,25 +96,25 @@ jsonrpc-core-client = { version = "18.0.0", optional = true, default-features = 
 jsonrpc-derive = { version = "18.0.0", optional = true, default-features = false }
 
 # manta dependencies
-manta-accounting = { git = "https://github.com/manta-network/manta-rs.git", branch = "feat/flexible-ledger-api", default-features = false }
-manta-crypto = { git = "https://github.com/manta-network/manta-rs.git", branch = "feat/flexible-ledger-api", default-features = false }
-manta-pay = { git = "https://github.com/manta-network/manta-rs.git", branch = "feat/flexible-ledger-api", default-features = false, features = ["groth16", "scale", "serde"] }
+manta-accounting = { git = "https://github.com/manta-network/manta-rs.git", default-features = false }
+manta-crypto = { git = "https://github.com/manta-network/manta-rs.git", default-features = false }
+manta-pay = { git = "https://github.com/manta-network/manta-rs.git", default-features = false, features = ["groth16", "scale"] }
 manta-sdk = { git = "https://github.com/manta-network/sdk.git", default-features = false }
-manta-util = { git = "https://github.com/manta-network/manta-rs.git", branch = "feat/flexible-ledger-api", default-features = false, features = ["serde", "serde_with"] }
-manta-primitives = { path = "../../primitives", default-features = false}
+manta-util = { git = "https://github.com/manta-network/manta-rs.git", default-features = false }
+manta-primitives = { path = "../../primitives", default-features = false }
 
 [dev-dependencies]
 bencher = "0.1.5"
 criterion = "0.3.4"
 lazy_static = "1.4.0"
-manta-accounting = { git = "https://github.com/manta-network/manta-rs.git", branch = "feat/flexible-ledger-api", features = ["test"] }
+manta-accounting = { git = "https://github.com/manta-network/manta-rs.git", features = ["test"] }
 manta-sdk = { git = "https://github.com/manta-network/sdk.git", features = ["download"] }
 pallet-asset-manager = { path = "../asset-manager"}
 pallet-assets = { git = 'https://github.com/paritytech/substrate.git', branch = "polkadot-v0.9.16" }
 pallet-balances = { git = 'https://github.com/paritytech/substrate.git', branch = "polkadot-v0.9.16" }
 rand = "0.8.4"
 sp-core = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.16"}
-sp-io = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.16"}
+sp-io = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.16" }
 tempfile = "3.3.0"
 xcm = { git = 'https://github.com/paritytech/polkadot.git', branch = "release-v0.9.16" }
