+++ pallets/manta-pay/src/lib.rs
@@ -71,7 +71,7 @@ use manta_crypto::{
 	constraint::ProofSystem,
 	merkle_tree::{self, forest::Configuration as _},
 };
-use manta_pay::config;
+use manta_pay::{config, signer::Checkpoint};
 use manta_primitives::{
 	assets::{AssetConfig, FungibleLedger as _, FungibleLedgerError},
 	types::{AssetId, Balance},
@@ -80,6 +80,9 @@ use manta_util::codec::Decode as _;
 use scale_codec::{Decode, Encode, MaxEncodedLen};
 use scale_info::TypeInfo;
 use types::*;
+
+#[cfg(feature = "serde")]
+use manta_util::serde::{Deserialize, Serialize};
 
 pub use pallet::*;
 pub use weights::WeightInfo;
@@ -93,11 +96,14 @@ mod test;
 pub mod types;
 pub mod weights;
 
+#[cfg(feature = "runtime-benchmarks")]
+pub mod benchmark;
+
 #[cfg(feature = "rpc")]
 pub mod rpc;
 
-#[cfg(feature = "runtime-benchmarks")]
-pub mod benchmark;
+#[cfg(feature = "runtime")]
+pub mod runtime;
 
 /// MantaPay Pallet
 #[frame_support::pallet]
@@ -448,6 +454,80 @@ pub mod pallet {
 	where
 		T: Config,
 	{
+		/// Maximum Number of Updates per Shard
+		const PULL_MAX_PER_SHARD_UPDATE_SIZE: usize = 128;
+
+		/// Maximum Size of Sender Data Update
+		const PULL_MAX_SENDER_UPDATE_SIZE: usize = 1024;
+
+		/// Pulls receiver data from the ledger starting at the `receiver_index`.
+		#[inline]
+		fn pull_receivers(receiver_index: &mut [usize; 256]) -> (bool, ReceiverChunk) {
+			let mut more_receivers = false;
+			let mut receivers = Vec::new();
+			for (i, index) in receiver_index.iter_mut().enumerate() {
+				more_receivers |= Self::pull_receivers_for_shard(i as u8, index, &mut receivers);
+			}
+			(more_receivers, receivers)
+		}
+
+		/// Pulls receiver data from the shard at `shard_index` starting at the `receiver_index`,
+		/// pushing the results back to `receivers`.
+		#[inline]
+		fn pull_receivers_for_shard(
+			shard_index: u8,
+			receiver_index: &mut usize,
+			receivers: &mut ReceiverChunk,
+		) -> bool {
+			let mut iter = if *receiver_index == 0 {
+				Shards::<T>::iter_prefix(shard_index)
+			} else {
+				let raw_key = Shards::<T>::hashed_key_for(shard_index, *receiver_index as u64 - 1);
+				Shards::<T>::iter_prefix_from(shard_index, raw_key)
+			};
+			for _ in 0..Self::PULL_MAX_PER_SHARD_UPDATE_SIZE {
+				match iter.next() {
+					Some((_, (utxo, encrypted_note))) => {
+						*receiver_index += 1;
+						receivers.push((utxo, encrypted_note));
+					}
+					_ => return false,
+				}
+			}
+			iter.next().is_some()
+		}
+
+		/// Pulls sender data from the ledger starting at the `sender_index`.
+		#[inline]
+		fn pull_senders(sender_index: &mut usize) -> (bool, SenderChunk) {
+			let mut senders = Vec::new();
+			let mut iter = VoidNumberSet::<T>::iter().skip(*sender_index);
+			for _ in 0..Self::PULL_MAX_SENDER_UPDATE_SIZE {
+				match iter.next() {
+					Some((sender, _)) => {
+						*sender_index += 1;
+						senders.push(sender);
+					}
+					_ => return (false, senders),
+				}
+			}
+			(iter.next().is_some(), senders)
+		}
+
+		/// Returns the update required to be synchronized with the ledger starting from
+		/// `checkpoint`.
+		#[inline]
+		pub fn pull(mut checkpoint: Checkpoint) -> PullResponse {
+			let (more_receivers, receivers) = Self::pull_receivers(&mut checkpoint.receiver_index);
+			let (more_senders, senders) = Self::pull_senders(&mut checkpoint.sender_index);
+			PullResponse {
+				should_continue: more_receivers || more_senders,
+				checkpoint,
+				receivers,
+				senders,
+			}
+		}
+
 		/// Returns the account ID of this pallet.
 		#[inline]
 		pub fn account_id() -> T::AccountId {
@@ -473,6 +553,39 @@ pub mod pallet {
 			Ok(().into())
 		}
 	}
+}
+
+/// Receiver Chunk Data Type
+pub type ReceiverChunk = Vec<(Utxo, EncryptedNote)>;
+
+/// Sender Chunk Data Type
+pub type SenderChunk = Vec<VoidNumber>;
+
+/// Ledger Source Pull Response
+#[cfg_attr(
+	feature = "serde",
+	derive(Deserialize, Serialize),
+	serde(crate = "manta_util::serde", deny_unknown_fields)
+)]
+#[derive(Clone, Debug, Decode, Default, Encode, Eq, Hash, PartialEq, TypeInfo)]
+pub struct PullResponse {
+	/// Pull Continuation Flag
+	///
+	/// The `should_continue` flag is set to `true` if the client should request more data from the
+	/// ledger to finish the pull.
+	pub should_continue: bool,
+
+	/// Ledger Checkpoint
+	///
+	/// If the `should_continue` flag is set to `true` then `checkpoint` is the next [`Checkpoint`]
+	/// to request data from the ledger. Otherwise, it represents the current ledger state.
+	pub checkpoint: Checkpoint,
+
+	/// Ledger Receiver Chunk
+	pub receivers: ReceiverChunk,
+
+	/// Ledger Sender Chunk
+	pub senders: SenderChunk,
 }
 
 /// Preprocessed Event
